requirements.txt
python-telegram-bot==20.0
openpyxl==3.1.2
